Precision upgrade! Welp we all know it's a potential breaking change, but I bet no one is using our API since we are still using 0.x.x as version number, semantic versioning complied

+++ src/main/java/snownee/cuisine/fluids/FluidDrink.java
@@ -200,9 +200,9 @@ public class FluidDrink extends VaporizableFluid
     public static class WeightedMaterial implements INBTSerializable<NBTTagCompound>
     {
         Material material;
-        float weight;
+        double weight;
 
-        public WeightedMaterial(Material material, float weight)
+        public WeightedMaterial(Material material, double weight)
         {
             this.material = material;
             this.weight = weight;
@@ -213,7 +213,7 @@ public class FluidDrink extends VaporizableFluid
         {
             NBTTagCompound data = new NBTTagCompound();
             data.setString(CuisineSharedSecrets.KEY_MATERIAL, material.getID());
-            data.setFloat("weight", weight);
+            data.setDouble("weight", weight);
             return data;
         }
